ShadeEditor: scale the player and caster
Scale can't yet be set to a custom value using the editor, but default
scale values are used to make the icons a reasonable size. The scale
values are stored in the exported file too.

caster.png and player.png are replaced by their original versions with
full resolution and transparent backgrounds.