ShadeEditor: Add trees with their shadows to levels
Repurpose Troll Stomp's Trees using one of our Plant assets. Render both
the plant and its shadow beneath it in the level editor. They're
arbitrarily scaled when rendered, but scale is still not configurable by
the user, and tree scale is not stored in the level file yet.

Remove Rocks, Bosses, and Minions.